chore(map-view): renaming directory
